# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TikkaMasalAI Backend",
    "version": "0.1.0"
  },
  "paths": {
    "/predict": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Predict Food Class",
        "description": "Run the image classification model on an uploaded image and return the top predicted labels with their probabilities.\n\nThis endpoint accepts a multipart/form-data POST with a file field named 'image'.",
        "operationId": "predict_endpoint_predict_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_predict_endpoint_predict_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Predict Endpoint Predict Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/predict/explain": {
      "post": {
        "tags": [
          "prediction"
        ],
        "summary": "Generate Attention Heatmap",
        "description": "Creates an attention-based visualization showing which parts of the image the model focuses on for prediction.",
        "operationId": "explain_prediction_predict_explain_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_explain_prediction_predict_explain_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Explain Prediction Predict Explain Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm/generate": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Generate text from prompt",
        "description": "Generate text with the configured Ollama model.\n\nPOST a JSON body with `prompt`, optional `temperature` and `max_tokens`.\nReturns the generated text and the model used.",
        "operationId": "generate_text_llm_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Text Llm Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/llm/health": {
      "get": {
        "tags": [
          "llm"
        ],
        "summary": "Ollama service health",
        "description": "Check whether the local Ollama service is reachable and return a list of available models plus the configured model name.\n\nReturns a JSON object with `status`, `ollama_url`, `available_models` and `configured_model`.",
        "operationId": "check_ollama_health_llm_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Check Ollama Health Llm Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Read Root",
        "operationId": "read_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Lightweight health endpoint for container orchestration.\n\nReturns 200 OK when the app is up and routers are registered.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_explain_prediction_predict_explain_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_explain_prediction_predict_explain_post"
      },
      "Body_predict_endpoint_predict_post": {
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_predict_endpoint_predict_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.7
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "default": 500
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "LLMRequest",
        "description": "Request model for LLM generation."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}